+++ crates/relayer-handlers/src/routes/mod.rs
@@ -1,3 +1,4 @@
+use ethereum_types::H512;
 use serde::{Deserialize, Serialize};
 
 /// Module for handling encrypted commitment leaves API
@@ -70,4 +71,33 @@ const fn default_zero() -> Option<u32> {
 
 const fn default_u32_max() -> Option<u32> {
     Some(u32::MAX)
+}
+
+/// Success response for withdrawal tx relaying API request.
+#[derive(Debug, Serialize)]
+#[serde(rename_all = "camelCase")]
+pub struct WithdrawTxSuccessResponse {
+    status: String,
+    message: String,
+    item_key: H512,
+}
+
+/// Failure response for withdrawal tx relaying API request.
+#[derive(Debug, Serialize)]
+#[serde(rename_all = "camelCase")]
+pub struct WithdrawTxFailureResponse {
+    status: String,
+    message: String,
+    reason: String,
+}
+
+/// Withdrawal tx relaying API request response
+#[derive(Debug, Serialize)]
+#[serde(untagged)]
+#[serde(rename_all = "camelCase")]
+pub enum WithdrawTxResponse {
+    /// Success response for withdrawal tx API request.
+    Success(WithdrawTxSuccessResponse),
+    /// Failure response for withdrawal tx API request.
+    Failure(WithdrawTxFailureResponse),
 }
